requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-environ==0.12.0
django-js-asset==3.1.2
django-query-counter==0.4.1
pillow==11.2.1
psycopg==3.2.9
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.13.2
